Fix deleting pull mirror in project repo settings
Pull mirrors were only being disabled when a user clicks the
project repo settings UI delete button.
This commit adds a pull mirror delete endpoint that correctly
destroys the database record.
It is important to correctly delete pull mirrors when the user
expects it as credentials are present in `import_data_attributes`.