Expose the namespace statistics
Add a statistics field on the namespace type in graphql

The statistics are loaded as namespaces in a separate
batchloader.

That means we are loading the same namespace twice: Once
for exposing the namespace itself, causing a query without summing up
the project statistics, and once with statistics for each namespace
we're rendering statistics for.