Change tier badge regex to use parenthesis
Previously, we used brackets to denote the tier badges,
but this made Kramdown, the docs site Markdown renderer,
to show some warnings when the site was built. This is
now changed by replacing brackets with parenthesis.

This is because [PREMIUM] is translated to a link, and Kramdown
wouldn't find its URL since there isn't one.

- https://gitlab.com/gitlab-com/gitlab-docs/merge_requests/484
- https://gitlab.com/gitlab-org/gitlab-ce/issues/63800